+++ packages/solid/src/components/tree-view/tree-view-node-checkbox.tsx
@@ -9,7 +9,7 @@ export interface TreeViewNodeCheckboxProps extends HTMLProps<'span'>, TreeViewNo
 export const TreeViewNodeCheckbox = (props: TreeViewNodeCheckboxProps) => {
   const treeView = useTreeViewContext()
   const nodeProps = useTreeViewNodePropsContext()
-  const mergedProps = mergeProps(() => treeView().getItemCheckboxProps(nodeProps), props)
+  const mergedProps = mergeProps(() => treeView().getNodeCheckboxProps(nodeProps), props)
 
   return <ark.span {...mergedProps} />
 }